configs: k3: remove saved environments
Having saved environments usually causes inconsistencies while
development workflow. The saved environments conflict with the
default ones that uboot should be updating while development
but that doesn't happen and the saved environments need to be
reset during bootups to test the changes causing extra debugs.

Remove the saved environments as a default.

( Uenv.txt can be used optionally as a replacement for saved
environments )

Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>